+++ src/engine/i_opndir.h
@@ -60,7 +60,7 @@
 #endif
 
 struct dirent {
-    long          d_ino;    /* Always zero. */
+    int          d_ino;    /* Always zero. */
     unsigned short d_reclen; /* Always zero. */
     unsigned short d_namlen; /* Length of name in d_name. */
     char           d_name[FILENAME_MAX]; /* File name. */
@@ -81,7 +81,7 @@ typedef struct {
     struct dirent dd_dir;
 
     /* _findnext handle */
-    long    dd_handle;
+    int    dd_handle;
 
     /*
      * Status of search:
@@ -99,8 +99,8 @@ DIR *opendir(const char *);
 struct dirent *readdir(DIR *);
 int closedir(DIR *);
 void rewinddir(DIR *);
-long telldir(DIR *);
-void seekdir(DIR *, long);
+int telldir(DIR *);
+void seekdir(DIR *, int);
 
 #endif
