+++ src/engine/m_misc.h
@@ -66,7 +66,7 @@ dboolean M_WriteFile(char const* name, void* source, int length);
 int M_ReadFile(char const* name, byte** buffer);
 void M_NormalizeSlashes(char *str);
 int M_FileExists(char *filename);
-long M_FileLength(FILE *handle);
+int M_FileLength(FILE *handle);
 dboolean M_WriteTextFile(char const* name, char* source, int length);
 void M_ScreenShot(void);
 int M_CacheThumbNail(byte** data);
